# backend/docs/swagger.yaml
basePath: /
definitions:
...
  dto.ContextRuleResponse:
    properties:
      createdAt:
        type: string
      flag:
        $ref: '#/definitions/dto.FlagObject'
      id:
        type: string
      keyword:
        type: string
      subflag:
        $ref: '#/definitions/dto.SubflagObject'
      updatedAt:
        type: string
    type: object
  dto.CreateContextRuleRequest:
    properties:
      flagId:
        type: string
      keyword:
        type: string
      subflagId:
        type: string
    type: object
  dto.CreateEventRequest:
    properties:
      allDay:
        type: boolean
      endAt:
        type: string
      flagId:
        type: string
      location:
        type: string
      startAt:
        type: string
      subflagId:
        type: string
      title:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
      requestId:
        type: string
    type: object
  dto.EventResponse:
    properties:
      allDay:
        type: boolean
      createdAt:
        type: string
      endAt:
        type: string
      flag:
        $ref: '#/definitions/dto.FlagObject'
      id:
        type: string
      location:
        type: string
      sourceInboxItem:
        $ref: '#/definitions/dto.InboxItemObject'
      startAt:
        type: string
      subflag:
        $ref: '#/definitions/dto.SubflagObject'
      title:
        type: string
      updatedAt:
        type: string
    type: object
  dto.FlagObject:
    properties:
      color:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  dto.InboxItemObject:
    properties:
      createdAt:
        type: string
      id:
        type: string
      lastError:
        type: string
      rawMediaUrl:
        type: string
      rawText:
        type: string
      source:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
...
  dto.ListContextRulesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.ContextRuleResponse'
        type: array
      nextCursor:
        type: string
    type: object
  dto.ListEventsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.EventResponse'
        type: array
      nextCursor:
        type: string
    type: object
...
  dto.RegisterTokenRequest:
    properties:
      appVersion:
        type: string
      deviceId:
        type: string
      deviceName:
        type: string
      platform:
        description: ios | android
        type: string
    required:
    - deviceId
    - platform
    type: object
  dto.RegisterTokenResponse:
    properties:
      topic:
        type: string
    type: object
...
  dto.SubflagObject:
    properties:
      color:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  dto.UnregisterTokenRequest:
    properties:
      deviceId:
        type: string
    required:
    - deviceId
    type: object
  dto.UpdateContextRuleRequest:
    properties:
      flagId:
        type: string
      keyword:
        type: string
      subflagId:
        type: string
    type: object
  dto.UpdateEventRequest:
    properties:
      allDay:
        type: boolean
      endAt:
        type: string
      flagId:
        type: string
      location:
        type: string
      startAt:
        type: string
      subflagId:
        type: string
      title:
        type: string
    type: object
...
info:
  contact: {}
  description: API do MVP Organiq.
  title: Organiq API
  version: "0.1"
paths:
...
  /v1/context-rules:
    get:
      parameters:
      - description: Limite
        in: query
        name: limit
        type: integer
      - description: Cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListContextRulesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar regras de contexto
      tags:
      - ContextRules
    post:
      consumes:
      - application/json
      parameters:
      - description: Context rule payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateContextRuleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ContextRuleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar regra de contexto
      tags:
      - ContextRules
  /v1/context-rules/{id}:
    delete:
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remover regra de contexto
      tags:
      - ContextRules
    patch:
      consumes:
      - application/json
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      - description: Context rule payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateContextRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ContextRuleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar regra de contexto
      tags:
      - ContextRules
...
  /v1/devices/token:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Unregister device request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UnregisterTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remover dispositivo
      tags:
      - Devices
    post:
      consumes:
      - application/json
      parameters:
      - description: Register device request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RegisterTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Registrar dispositivo
      tags:
      - Devices
  /v1/events:
    get:
      parameters:
      - description: Limite
        in: query
        name: limit
        type: integer
      - description: Cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListEventsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar eventos
      tags:
      - Events
    post:
      consumes:
      - application/json
      parameters:
      - description: Event payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.EventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar evento
      tags:
      - Events
  /v1/events/{id}:
    delete:
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Excluir evento
      tags:
      - Events
    patch:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Event payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar evento
      tags:
      - Events
...
securityDefinitions:
  BearerAuth:
    description: 'Formato: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"